+++ src/types/node-transformers/string-array-transformers/TStringArrayScopeCallsWrappersDataByEncoding.ts
@@ -1,0 +1,7 @@
+import { TStringArrayEncoding } from '../../options/TStringArrayEncoding';
+
+import { IStringArrayScopeCallsWrappersData } from '../../../interfaces/node-transformers/string-array-transformers/IStringArrayScopeCallsWrappersData';
+
+export type TStringArrayScopeCallsWrappersDataByEncoding = Partial<{
+    [key in TStringArrayEncoding]: IStringArrayScopeCallsWrappersData;
+}>;